+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- `requiresConfirmation` to `FormElement`
+
 ## [11.0.4] - 2025-05-22
 
 ### Added
 
-- `requiresConfirmation` to `FormElement`
 - `googleAnalyticsTagId` to `FormsAppBase`
 - Additional attributes to `RawJWTPayload`
 - Additional attributes to `FormSubmissionUser`